db: patch cksumvfs code imported from the PoC
Review fixes to the cksumvfs code imported verbatim from the PoC, kept
separate so that import stays unmodified.

- needsCksumVacuum: accept only 0 (legacy) or 8 (migrated) reserve bytes
  and report anything else as ErrDBCorrupted.
- Activate cksumvfs in New without the connect hook: run the reserve-byte
  file-control and VACUUM back to back on one connection and reopen
  afterwards, as cksumvfs requires, before creating tables.
- Drop the misleading PRAGMA integrity_check mention from a test comment.
- Move the cksumvfs tests into fin_test.go and delete cksumvfs_test.go.
